API reference

FiniteDifferenceMatrices.Fornberg1988

FiniteDifferenceMatrices.fdcoefficient

FiniteDifferenceMatrices.fdmatrix

FiniteDifferenceMatrices.fdvalue

FiniteDifferenceMatrices.Fornberg1988 — Constant

Fornberg1988[d,n,m]

This is a foolish implementation of the tables of B. Fornberg, Math. Comp. 51 699-706 (1988) and Wikipedia by hand. It was coded to test the fdcoefficient().

Examples

julia> FiniteDifferenceMatrices.Fornberg1988[:c,1,2]
1×3 Matrix{Rational{Int64}}:
 -1//2  0//1  1//2

julia> FiniteDifferenceMatrices.Fornberg1988[:c,2,2]
1×3 Matrix{Rational{Int64}}:
 1//1  -2//1  1//1

FiniteDifferenceMatrices.fdcoefficient — Method

fdcoefficient(; n::Int=1, m::Int=2, d=:c, t=Rational{Int})

This function returns a Dict of the finite difference coefficients $c_i$ of

\[\frac{\mathrm{d}^n f}{\mathrm{d}x^n}(x) = \frac{1}{\Delta x^n} \sum_{i} c_i f(x+i\Delta x) + O(\Delta x^m).\]

This implementation is based on a post on discourse by @stevengj and this function is tested to return results equivalent to B. Fornberg, Math. Comp. 51 699-706 (1988).

Arguments	Description

n	order of derivative $n$

m	order of accuracy $m$

d	direction, :c central, :f forward, :b backward

t	type of coefficients, e.g.: Rational{Int}, Rational{BigInt}, Rational{Float64}

Examples

The coefficients of the central, $n=1$ and $m=2$ differences are $c_{-1} = -1/2, c_{0} = 0, c_{1} = 1/2$.

\[\frac{\mathrm{d}f}{\mathrm{d} x}(x) = \frac{f(x+\Delta x) - f(x-\Delta x)}{2\Delta x} + O(\Delta x^{2})\]

julia> fdcoefficient(n=1, m=2, d=:c)
Dict{Int64, Rational{Int64}} with 3 entries:
  0  => 0//1
  -1 => -1//2
  1  => 1//2

The coefficients of the central, $n=1$ and $m=1$ differences are $c_{0} = -1, c_{1} = 1$.

\[\frac{\mathrm{d}f}{\mathrm{d} x}(x) = \frac{f(x+\Delta x) - f(x)}{\Delta x} + O(\Delta x)\]

julia> fdcoefficient(n=1, m=1, d=:f)
Dict{Int64, Rational{Int64}} with 2 entries:
  0 => -1//1
  1 => 1//1

The coefficients of the central, $n=2$ and $m=2$ differences are $c_{-1} = 1, c_{0} = -2, c_{1} = 1$.

\[\frac{\mathrm{d}^{2}f}{\mathrm{d} x^{2}}(x) = \frac{f(x+\Delta x) - 2f(x) + f(x-\Delta x)}{\Delta x^{2}} + O(\Delta x^{2})\]

julia> fdcoefficient(n=2, m=2, d=:c)
Dict{Int64, Rational{Int64}} with 3 entries:
  0  => -2//1
  -1 => 1//1
  1  => 1//1

FiniteDifferenceMatrices.fdmatrix — Method

fdmatrix(N::Int; n::Int=1, m::Int=2, d=:c, h=0.1, t=Rational{Int})

This function returns a discrete approximation of a differential operator as a SparseMatrixCSC. The numerical error is not ignored in high orders ($3<n$, $4<m$) and a small grid spacing ($h<10^{-3}$). Please use high precision arithmetic (h=big"0.001" and t=Rational{BigInt}) in that case.

Arguments	Description

n	order of derivative $n$

m	order of accuracy $m$

d	direction, :c central, :f forward, :b backward

h	grid spacing $\Delta x$

t	type of coefficients, e.g.: Rational{Int}, Rational{BigInt}, Float64

Examples

The central, $n=1$ and $m=2$ discrete approximation of the differential operator is

\[\frac{\mathrm{d}}{\mathrm{d} x} \simeq \frac{1}{2\Delta x} \left(\begin{array}{ccccccc} 0 & 1 & 0 & \ldots & 0 & 0 & 0 \\ -1 & 0 & 1 & \ldots & 0 & 0 & 0 \\ 0 & -1 & 0 & \ldots & 0 & 0 & 0 \\ \vdots & \vdots & \vdots & \ddots & \vdots & \vdots & \vdots \\ 0 & 0 & 0 & \ldots & 0 & 1 & 0 \\ 0 & 0 & 0 & \ldots & -1 & 0 & 1 \\ 0 & 0 & 0 & \ldots & 0 & -1 & 0 \end{array}\right).\]

julia> fdmatrix(5, n=1, m=2, d=:c, h=1//1)
5×5 SparseArrays.SparseMatrixCSC{Rational{Int64}, Int64} with 8 stored entries:
   ⋅     1//2    ⋅      ⋅     ⋅  
 -1//2    ⋅     1//2    ⋅     ⋅  
   ⋅    -1//2    ⋅     1//2   ⋅  
   ⋅      ⋅    -1//2    ⋅    1//2
   ⋅      ⋅      ⋅    -1//2   ⋅  

The central, $n=1$ and $m=1$ discrete approximation of the differential operator is

\[\frac{\mathrm{d}}{\mathrm{d} x} \simeq \frac{1}{\Delta x} \left(\begin{array}{ccccccc} -1 & 1 & 0 & \ldots & 0 & 0 & 0 \\ 0 & -1 & 1 & \ldots & 0 & 0 & 0 \\ 0 & 0 & -1 & \ldots & 0 & 0 & 0 \\ \vdots & \vdots & \vdots & \ddots & \vdots & \vdots & \vdots \\ 0 & 0 & 0 & \ldots & -1 & 1 & 0 \\ 0 & 0 & 0 & \ldots & 0 & -1 & 1 \\ 0 & 0 & 0 & \ldots & 0 & 0 & -1 \end{array}\right).\]

julia> fdmatrix(5, n=1, m=1, d=:f, h=1//1)
5×5 SparseArrays.SparseMatrixCSC{Rational{Int64}, Int64} with 9 stored entries:
 -1//1   1//1    ⋅      ⋅      ⋅  
   ⋅    -1//1   1//1    ⋅      ⋅  
   ⋅      ⋅    -1//1   1//1    ⋅  
   ⋅      ⋅      ⋅    -1//1   1//1
   ⋅      ⋅      ⋅      ⋅    -1//1

The central, $n=2$ and $m=2$ discrete approximation of the differential operator is

\[\frac{\mathrm{d}^2}{\mathrm{d} x^2} \simeq \frac{1}{\Delta x^2} \left(\begin{array}{ccccccc} -2 & 1 & 0 & \ldots & 0 & 0 & 0 \\ 1 & -2 & 1 & \ldots & 0 & 0 & 0 \\ 0 & 1 & -2 & \ldots & 0 & 0 & 0 \\ \vdots & \vdots & \vdots & \ddots & \vdots & \vdots & \vdots \\ 0 & 0 & 0 & \ldots & -2 & 1 & 0 \\ 0 & 0 & 0 & \ldots & 1 & -2 & 1 \\ 0 & 0 & 0 & \ldots & 0 & 1 & -2 \end{array}\right).\]

julia> fdmatrix(5, n=2, m=2, d=:c, h=1//1)
5×5 SparseArrays.SparseMatrixCSC{Rational{Int64}, Int64} with 13 stored entries:
 -2//1   1//1    ⋅      ⋅      ⋅  
  1//1  -2//1   1//1    ⋅      ⋅  
   ⋅     1//1  -2//1   1//1    ⋅  
   ⋅      ⋅     1//1  -2//1   1//1
   ⋅      ⋅      ⋅     1//1  -2//1

FiniteDifferenceMatrices.fdvalue — Method

fdvalue(f, a; n::Int=1, m::Int=2, d=:c, h=0.1, t=Rational{Int})

This function calculates the differential coefficient $f^{(n)}(a)$, a value of the derivative function $f^{(n)}=\frac{\mathrm{d}^n f}{\mathrm{d}x^n}$ at the point $a$.

\[\frac{\mathrm{d}^n f}{\mathrm{d}x^n} (a) = \frac{1}{h^n} \sum_{i} c_i f(a+ih) + O(\Delta x^m).\]

The numerical error is not ignored in high orders ($3<n$, $4<m$) and a small grid spacing ($h<10^{-3}$). Please use high precision arithmetic (h=big"0.001" and t=Rational{BigInt}) in that case.

Examples

julia> fdvalue(x -> x^2, 1.0)
2.0000000000000004

julia> (x -> 2*x)(1.0)
2.0

julia> fdvalue(sin, 0.0)
0.9983341664682815

julia> fdvalue(sin, 0.0, m=4)
0.9999966706326067

julia> fdvalue(sin, 0.0, m=6)
0.9999999928710186

julia> cos(0.0)
1.0